Fix bugs and clean up the API codebase
Bugs fixed:
- Stop swallowing exceptions: actions no longer return an empty Ok() on
  failure; add a global exception handler that returns a proper 500
  response with the standard envelope
- Add [Authorize(Roles=...)] to MakeAdmin, MakeOwner, RemoveAdminRole and
  RemoveOwnerRole so users can no longer self-promote
- Fix AuthController routes being doubled by combining a class-level
  [Route] with full-path action routes (api/auth/api/auth/...)
- Call base.OnModelCreating() in AppDbContext so the Identity schema is
  configured; remove the redundant HasNoKey() overrides
- Use double instead of float for Latitude/Longitude to avoid coordinate
  precision loss
- Make duplicate-name checks case-insensitive
- Add unique indexes on Region/District/Town names to prevent duplicates
  at the DB level
- Fix RegisterAsync checking username via FindByEmailAsync; check email
  and username separately
- Return 200 from LoginAsync instead of 201; use UTC for JWT expiry
- AuthController now maps the service status code to the real HTTP status

Cleanup:
- Switch controllers to ControllerBase and drop redundant try/catch,
  null and ModelState checks (handled by [ApiController])
- Remove unused MongoDB.Driver and NewtonsoftJson packages; use a single
  System.Text.Json serializer with ReferenceHandler.IgnoreCycles
- Remove Save() from repository interfaces, stop copying entities on
  create, use AsNoTracking() for reads, and skip wasted saves on update
  when the entity is not found
- Normalize API routes in ApiEndpoints (no more /region/region/{id} style
  segments; lowercase auth base)
- Add validation attributes to all request DTOs

Build:
- Add InitialCreate migration (Identity tables, unique indexes, double
  precision coordinates)

+++ AccraCityApi.Contracts/Response/TownResponses/TownResponse.cs
@@ -7,8 +7,8 @@ public class TownResponse
     public required string TownName { get; set; }
     public required string Category { get; set; }
     public required int Population { get; set; }
-    public required float Latitude { get; set; }
-    public required float Longitude { get; set; }
+    public required double Latitude { get; set; }
+    public required double Longitude { get; set; }
     public DateTime CreatedAt { get; set; }
     public DateTime LastModifiedAt { get; set; }
     public List<string> NearbyTowns { get; set; } = new List<string>();
